merge master

+++ ARMS-Backend/src/main/java/ca/mcgill/ecse321/arms/service/AppointmentService.java
@@ -1,11 +1,9 @@
 package ca.mcgill.ecse321.arms.service;
 
 import ca.mcgill.ecse321.arms.ArmsApplication;
-import ca.mcgill.ecse321.arms.dao.AppointmentRepository;
-import ca.mcgill.ecse321.arms.dao.CarRepository;
-import ca.mcgill.ecse321.arms.dao.ServiceRepository;
-import ca.mcgill.ecse321.arms.dao.TimeSlotRepository;
+import ca.mcgill.ecse321.arms.dao.*;
 import ca.mcgill.ecse321.arms.model.Appointment;
+import ca.mcgill.ecse321.arms.service.TimeSlotService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -14,8 +12,9 @@ import ca.mcgill.ecse321.arms.model.*;
 
 import java.sql.Date;
 import java.sql.Time;
-import java.util.ArrayList;
-import java.util.List;
+import java.util.*;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 @Service
 public class AppointmentService {
@@ -27,38 +26,60 @@ public class AppointmentService {
     private TimeSlotRepository timeSlotRepository;
     @Autowired
     private CarRepository carRepository;
+    @Autowired
+    private BusinessHourRepository businessHourRepository;
+    @Autowired
+    private BussinessRepository bussinessRepository;
+    @Autowired
+    private SpaceRepository spaceRepository;
+    @Autowired
+    private TechnicianRepository technicianRepository;
 
     /**
      * add an appointment to the ARMS
      *
-     * @param service  service of the appointment
-     * @param car      car of the appointment
-     * @param timeSlot time slot of the appointment
+     * @param serviceName  service name of the appointment
+     * @param plateNo      car plate number of the appointment
+     * @param businessName business name of the appointment
+     * @param startDate
+     * @param startTime
+     * @param endDate
+     * @param endTime
+     * @param spaceID
+     * @param technicianID
      * @return the appointment that is added to the ARMS
      * @throws IllegalArgumentException
      * @author Grey Yuan
      */
     @Transactional
-    public Appointment createAppointment(ca.mcgill.ecse321.arms.model.Service service, Car car, TimeSlot timeSlot) {
+    public Appointment createAppointment(String serviceName, String plateNo,
+                                         String businessName,String startDate, String startTime,
+                                         String endDate, String endTime,int spaceID,int technicianID) {
         String error = "";
         // parameter check
-        if (service.toString().isEmpty()) {
-            error += "No service is associated with the appointment\n";
+        if(serviceName.equals("")){
+            error += "You must enter a service name\n";
         }
-        if (car.toString().isEmpty()) {
-            error += "No car is associated with the appointment\n";
+        if (plateNo.equals("")) {
+            error += "You must enter a plate number\n";
         }
-        if (timeSlot.toString().isEmpty()) {
-            error += "No time slot is associated with the appointment";
+        if (businessName==null || startDate==null || startTime==null || endDate==null || endTime==null){
+            error = "The time slot parameter cannot be empty";
+        }
+        if (serviceRepository.findServiceByName(serviceName) == null) {
+            error = "The service does not exist";
+        }
+        if (carRepository.findCarByPlateNo(plateNo) == null) {
+            error = "The car does not exist";
         }
         if (error.length() > 0) throw new IllegalArgumentException(error);
 
-        // if the appointment of the customer already exists
-        Appointment appointment = appointmentRepository.findAppointmentByCar(car);
-        if (service != null) {
-            error = "appointment of " + car.getCustomer().toString() + " already exists";
-            throw new IllegalArgumentException(error);
-        }
+        //find the intermediate objects with given paremeter
+        ca.mcgill.ecse321.arms.model.Service service = serviceRepository.findServiceByName(serviceName);
+        Car car = carRepository.findCarByPlateNo(plateNo);
+        TimeSlot timeSlot = createTimeSlot(businessName,startDate,startTime,endDate,endTime,spaceID,technicianID);
+
+
 
         Appointment anAppointment = new Appointment();
         anAppointment.setService(service);
@@ -76,29 +97,46 @@ public class AppointmentService {
      * @author Grey Yuan
      */
     @Transactional
-    public void deleteAppointment(int appointmentID) {
+    public int deleteAppointment(int appointmentID) {
+        String error = "";
         Appointment anAppointment = appointmentRepository.findAppointmentByAppointmentID(appointmentID);
         if(anAppointment==null){
             throw new IllegalArgumentException("No appointment was found.");
         }
+        //only be able to cancel 24hr before the appointment starts
+        Date appointmentDate = anAppointment.getTimeSlot().getStartDate();
+        Date curDate = ArmsApplication.getCurrentDate();
+        if (appointmentDate.equals(curDate)){
+            error = "You can not delete an appointment happening on the current date";
+            throw new IllegalArgumentException(error);
+        }
 
         appointmentRepository.delete(anAppointment);
+        return appointmentID;
     }
 
     /**
      * update an appointment by indicating the appointment ID
      * @param appointmentID the appointment that will be updated
-     * @param service new service of the appointment
-     * @param car new car
-     * @param timeSlot new time slot
+     * @param serviceName  service name of the appointment
+     * @param plateNo      car plate number of the appointment
+     * @param businessName business name of the appointment
+     * @param startDate
+     * @param startTime
+     * @param endDate
+     * @param endTime
+     * @param spaceID
+     * @param technicianID
      * @throws IllegalArgumentException
      * @return the appointment that is updated
      * @author Grey Yuan
      */
     @Transactional
-    public Appointment updateService(int appointmentID, ca.mcgill.ecse321.arms.model.Service service, Car car, TimeSlot timeSlot){
+    public Appointment updateAppointment(int appointmentID, String serviceName, String plateNo,
+                                     String businessName,String startDate, String startTime,
+                                     String endDate, String endTime,int spaceID,int technicianID){
         deleteAppointment(appointmentID);
-        return createAppointment(service,car,timeSlot);
+        return createAppointment(serviceName,plateNo,businessName,startDate,startTime,endDate,endTime,spaceID,technicianID);
     }
 
     /**
@@ -111,7 +149,7 @@ public class AppointmentService {
     public Appointment getAppointment(int appointmentID){
         Appointment appointment = appointmentRepository.findAppointmentByAppointmentID(appointmentID);
         if(appointment==null){
-            throw new IllegalArgumentException("Appointment with ID" + appointmentID + " does not exist");
+            throw new IllegalArgumentException("Appointment with ID " + appointmentID + " does not exist");
         }
         return appointment;
     }
@@ -132,4 +170,110 @@ public class AppointmentService {
         }
         return resultList;
     }
+
+    public TimeSlot createTimeSlot(String businessName,String startDate1, String startTime1, String endDate1, String endTime1,int spaceID,int technicianID){
+
+        Date startDate = Date.valueOf(startDate1);
+        Time startTime = Time.valueOf(startTime1);
+        Date endDate = Date.valueOf(endDate1);
+        Time endTime = Time.valueOf(endTime1);
+
+        Business business = bussinessRepository.findBusinessByName(businessName);
+        Space space = spaceRepository.findSpaceBySpaceID(spaceID);
+        Technician technician = technicianRepository.findTechnicianByTechnicianID(technicianID);
+
+        System.out.println("Hi*2");
+        //Judge if has conflict with businesshour
+        List<BusinessHour> list_businessHour = businessHourRepository.findBusinessHourByBusiness(business);
+        Stream<BusinessHour> sorted1 = list_businessHour.stream().sorted(Comparator.comparing(BusinessHour::getBusinessHourID));
+        List<BusinessHour> list_businessHour_sorted = sorted1.collect(Collectors.toList());
+
+        System.out.println(list_businessHour_sorted.get(0).getBusinessHourID()+"and "+list_businessHour_sorted.get(list_businessHour_sorted.size()-1).getBusinessHourID());
+        //O(n)  comp
+        int flag1 = check_hour(list_businessHour_sorted,startDate,startTime,endDate,endTime);
+
+        if(flag1 != 0){
+            throw new IllegalArgumentException("cannot build such appointment since no free businessHour!");
+        }
+
+        //Judge if has conflict with space and tech
+        List<TimeSlot> list_timeSlot_Space = timeSlotRepository.findTimeSlotsBySpace(space);
+        List<TimeSlot> list_timeSlot_Tech = timeSlotRepository.findTimeSlotsByTechnician(technician);
+
+        Stream<TimeSlot> sorted2 = list_timeSlot_Space.stream().sorted(Comparator.comparing(TimeSlot::getTimeslotID));
+        List<TimeSlot> list_timeSlot_Space_sorted = sorted2.collect(Collectors.toList());
+
+        Stream<TimeSlot> sorted3 = list_timeSlot_Tech.stream().sorted(Comparator.comparing(TimeSlot::getTimeslotID));
+        List<TimeSlot> list_timeSlot_Tech_sorted = sorted3.collect(Collectors.toList());
+
+        int flag2 = check_slot(list_timeSlot_Space_sorted,startDate,startTime,endDate,endTime);
+        int flag3 = check_slot(list_timeSlot_Tech_sorted,startDate,startTime,endDate,endTime);
+
+        if(flag2 != 0){
+            throw new IllegalArgumentException("cannot build such appointment since no free space!");
+        }
+        if(flag3 != 0){
+            throw new IllegalArgumentException("cannot build such appointment since no free tech !");
+        }
+
+        TimeSlot timeSlot = new TimeSlot();
+        timeSlot.setStartDate(startDate);
+        timeSlot.setStartTime(startTime);
+        timeSlot.setEndDate(endDate);
+        timeSlot.setEndTime(endTime);
+        timeSlot.setTimeslotID(transfer(startDate, startTime)*100+spaceID*10+technicianID);
+        timeSlot.setSpace(space);
+        timeSlot.setTechnician(technician);
+        System.out.println("itmeSlotID is "+timeSlot.getTimeslotID());
+
+        timeSlotRepository.save(timeSlot);
+        return timeSlot;
+    }
+
+    private int check_hour(List<BusinessHour> list_businessHour,Date startDate,Time startTime,Date endDate,Time endTime){
+        //O(n)  comp
+        int flag1 = 0; // if has conflict, flag1 = 1
+        int i = 0;
+        int q = 0;
+        for(;i<list_businessHour.size();i++){
+            if (!list_businessHour.get(i).getStartDate().after(startDate) && !list_businessHour.get(i).getEndDate().before(endDate)){
+                q=1;
+                System.out.println("i is "+i+"id is "+list_businessHour.get(i).getBusinessHourID());
+                if(list_businessHour.get(i).getEndTime().before(endTime) || list_businessHour.get(i).getStartTime().after(startTime)){
+                    System.out.println(list_businessHour.get(i).getEndTime().before(endTime)+" and "+list_businessHour.get(i).getStartTime().after(startTime));
+                    flag1 = 1;
+                    return flag1;
+                }
+            }
+        }
+        if(q == 0){ flag1 = 1;}
+        return flag1;
+    }
+
+    private int check_slot(List<TimeSlot> list_slot,Date startDate,Time startTime,Date endDate,Time endTime){
+        //O(n)  comp
+        int flag2 = 0; // if has conflict, flag1 = 1
+        int i = 0;
+        for(;i<list_slot.size();i++){
+            if (list_slot.get(i).getStartDate().equals(startDate)){
+                if(list_slot.get(i).getEndTime().before(startTime) || list_slot.get(i).getStartTime().after(endTime)){
+                    flag2 = 0;
+                }
+                else {
+                    flag2 = 1;
+                    return flag2;
+                }
+            }
+        }
+        return flag2;
+    }
+
+    private Long transfer(Date startDate, Time startTime){
+        String date = startDate.toString();
+        String time = startTime.toString();
+        String res = date+time;
+        res = res.replaceAll("[^a-zA-Z0-9\\u4E00-\\u9FA5]", "");
+        return Long.parseLong(res);
+    }
+
 }
